refactor: use unified checksum computation helpers
This commit unifies the checksum computation for different layers and splits it into a computation and folding step to make it more flexible.
