+++ layers/ip4.go
@@ -149,33 +149,15 @@ func (ip *IPv4) SerializeTo(b gopacket.SerializeBuffer, opts gopacket.SerializeO
 	}
 
 	if opts.ComputeChecksums {
-		ip.Checksum = checksum(bytes)
+		// Clear checksum bytes
+		bytes[10] = 0
+		bytes[11] = 0
+
+		csum := gopacket.ComputeChecksum(bytes, 0)
+		ip.Checksum = gopacket.FoldChecksum(csum)
 	}
 	binary.BigEndian.PutUint16(bytes[10:], ip.Checksum)
 	return nil
-}
-
-func checksum(bytes []byte) uint16 {
-	// Clear checksum bytes
-	bytes[10] = 0
-	bytes[11] = 0
-
-	// Compute checksum
-	var csum uint32
-	for i := 0; i < len(bytes); i += 2 {
-		csum += uint32(bytes[i]) << 8
-		csum += uint32(bytes[i+1])
-	}
-	for {
-		// Break when sum is less or equals to 0xFFFF
-		if csum <= 65535 {
-			break
-		}
-		// Add carry to the sum
-		csum = (csum >> 16) + uint32(uint16(csum))
-	}
-	// Flip all the bits
-	return ^uint16(csum)
 }
 
 func (ip *IPv4) flagsfrags() (ff uint16) {
