+++ layers/ip4_test.go
@@ -10,7 +10,6 @@ package layers
 
 import (
 	"bytes"
-	"encoding/binary"
 	"encoding/hex"
 	"net"
 	"reflect"
@@ -98,38 +97,6 @@ func serialize(ip *IPv4) ([]byte, error) {
 		ComputeChecksums: true,
 	})
 	return buffer.Bytes(), err
-}
-
-// Test the function checksum
-func TestChecksum(t *testing.T) {
-	testData := []struct {
-		name   string
-		header string
-		want   string
-	}{{
-		name:   "sum has two carries",
-		header: "4540005800000000ff11ffff0aeb1d070aed8877",
-		want:   "fffe",
-	}, {
-		name:   "wikipedia case",
-		header: "45000073000040004011b861c0a80001c0a800c7",
-		want:   "b861",
-	}}
-
-	for _, test := range testData {
-		bytes, err := hex.DecodeString(test.header)
-		if err != nil {
-			t.Fatalf("Failed to Decode header: %v", err)
-		}
-		wantBytes, err := hex.DecodeString(test.want)
-		if err != nil {
-			t.Fatalf("Failed to decode want checksum: %v", err)
-		}
-
-		if got, want := checksum(bytes), binary.BigEndian.Uint16(wantBytes); got != want {
-			t.Errorf("In test %q, got incorrect checksum: got(%x), want(%x)", test.name, got, want)
-		}
-	}
 }
 
 func TestIPv4InvalidOptionLength(t *testing.T) {
