+++ layers/tcpip.go
@@ -47,32 +47,11 @@ func (ip *IPv6) pseudoheaderChecksum() (csum uint32, err error) {
 	return csum, nil
 }
 
-// Calculate the TCP/IP checksum defined in rfc1071.  The passed-in csum is any
-// initial checksum data that's already been computed.
-func tcpipChecksum(data []byte, csum uint32) uint16 {
-	// to handle odd lengths, we loop to length - 1, incrementing by 2, then
-	// handle the last byte specifically by checking against the original
-	// length.
-	length := len(data) - 1
-	for i := 0; i < length; i += 2 {
-		// For our test packet, doing this manually is about 25% faster
-		// (740 ns vs. 1000ns) than doing it by calling binary.BigEndian.Uint16.
-		csum += uint32(data[i]) << 8
-		csum += uint32(data[i+1])
-	}
-	if len(data)%2 == 1 {
-		csum += uint32(data[length]) << 8
-	}
-	for csum > 0xffff {
-		csum = (csum >> 16) + (csum & 0xffff)
-	}
-	return ^uint16(csum)
-}
-
 // computeChecksum computes a TCP or UDP checksum.  headerAndPayload is the
 // serialized TCP or UDP header plus its payload, with the checksum zero'd
 // out. headerProtocol is the IP protocol number of the upper-layer header.
-func (c *tcpipchecksum) computeChecksum(headerAndPayload []byte, headerProtocol IPProtocol) (uint16, error) {
+// The returned 32bit checksum may need to be folded.
+func (c *tcpipchecksum) computeChecksum(headerAndPayload []byte, headerProtocol IPProtocol) (uint32, error) {
 	if c.pseudoheader == nil {
 		return 0, errors.New("TCP/IP layer 4 checksum cannot be computed without network layer... call SetNetworkLayerForChecksum to set which layer to use")
 	}
@@ -84,7 +63,9 @@ func (c *tcpipchecksum) computeChecksum(headerAndPayload []byte, headerProtocol 
 	csum += uint32(headerProtocol)
 	csum += length & 0xffff
 	csum += length >> 16
-	return tcpipChecksum(headerAndPayload, csum), nil
+
+	csum = gopacket.ComputeChecksum(headerAndPayload, csum)
+	return csum, nil
 }
 
 // SetNetworkLayerForChecksum tells this layer which network layer is wrapping it.
